MAINT: Move conversion of indexing for arrays to indexing module.
