+++ src/darsia/image/indexing.py
@@ -24,6 +24,8 @@ the x-axis.
 
 from typing import Union
 
+import numpy as np
+
 
 def to_matrix_indexing(axis: Union[str, int], indexing: str) -> str:
     """Conversion of single axis in Cartesian to matrix indexing.
@@ -169,3 +171,58 @@ def interpret_indexing(axis: str, indexing: str) -> tuple[int, bool]:
     # If the method reaches this point, something went wrong.
     # This fact is used as safety check.
     raise ValueError
+
+
+def matrixToCartesianIndexing(img: np.ndarray) -> np.ndarray:
+    """
+    Reordering data indexing converting from (row,col) to (x,y) indexing.
+
+    The conventional matrix indexing uses the (row, col) format, such
+    that the top left corner is the (0,0) pixel. On the other hand,
+    Cartesian indexing uses the (x,y) order and thereby identifies
+    the lower left corner by (0,0). This routine is in particular useful
+    when communicating image data to conventional simulators, which use
+    the Cartesian indexing.
+
+    NOTE: Assumes 2d images.
+
+    Arguments:
+        np.ndarray: image array with matrix indexing
+
+    Returns:
+        np.ndarray: image array with Cartesian indexing
+    """
+    # Two operations are require: Swapping axis and flipping the vertical axis.
+
+    # Exchange first and second component, to change from (row,col) to (x,y) format.
+    img = np.swapaxes(img, 0, 1)
+
+    # Flip the orientation of the second axis, such that later y=0 is located at the bottom.
+    img = np.flip(img, 1)
+
+    return img
+
+
+def cartesianToMatrixIndexing(img: np.ndarray) -> np.ndarray:
+    """
+    Reordering data indexing, converting from (x,y) to (row,col) indexing.
+
+    Inverse to matrixToCartesianIndexing.
+
+    NOTE: Assumes 2d images.
+
+    Arguments:
+        np.ndarray: image array with Cartesian indexing
+
+    Returns:
+        np.ndarray: image array with matrix indexing
+    """
+    # Two operations are require: Swapping axis and flipping the vertical axis.
+
+    # Flip the orientation of the second axis, such that later row=0 is located at the top.
+    img = np.flip(img, 1)
+
+    # Exchange first and second component, to change from (x,y) to (row,col) format.
+    img = np.swapaxes(img, 0, 1)
+
+    return img
